spark-stream.py
from cassandra.cluster import Cluster
from pyspark.sql import SparkSession
from pyspark.sql.functions import from_json, col
from pyspark.sql.types import StructType, StructField, StringType

def get_cassandra_connection():
    try:
        cluster = Cluster(['localhost'])
        cas_session = cluster.connect()
        return cas_session
    except Exception as e:
        print(f"[ERROR] Could not create cassandra connection due to {e}")
        return None
    
def create_cassandra_keyspace(session):
    session.execute("""
        CREATE KEYSPACE IF NOT EXISTS spark_stream
        WITH replication = {'class': 'SimpleStrategy', 'replication_factor': '1'};
    """)

    print("[INFO] Cassandra keyspace created successfully!")

def create_cassandra_table(session):
    session.execute("""
    CREATE TABLE IF NOT EXISTS spark_stream.users (
        id UUID PRIMARY KEY,
        first_name TEXT,
        last_name TEXT,
        gender TEXT,
        address TEXT,
        post_code TEXT,
        email TEXT,
        username TEXT,
        registered_date TEXT,
        phone TEXT,
        picture TEXT);
    """)

    print("[INFO] Cassandra table created successfully!")

def get_spark_connection():
    s_conn = None

    try:
        s_conn = SparkSession.builder \
            .appName('SparkDataStreaming') \
            .config('spark.jars.packages', "com.datastax.spark:spark-cassandra-connector_2.13:3.5.1,"
                                           "org.apache.spark:spark-sql-kafka-0-10_2.13:3.5.7") \
            .config('spark.cassandra.connection.host', 'localhost') \
            .getOrCreate()

        print("[INFO] Spark connection created successfully!")
    except Exception as e:
        print(f"[ERROR] Couldn't create the spark session due to exception {e}")

    return s_conn

def get_data_from_kafka(spark_conn):
    spark_df = None
    try:
        spark_df = spark_conn.readStream \
            .format('kafka') \
            .option('kafka.bootstrap.servers', 'localhost:9092') \
            .option('subscribe', 'users') \
            .option('startingOffsets', 'earliest') \
            .load()
    except Exception as e:
        print(f"[ERROR] kafka dataframe could not be created because: {e}")

    return spark_df

def parse_kafka_stream(spark_df):
    schema = StructType([
        StructField("id", StringType(), False),
        StructField("first_name", StringType(), False),
        StructField("last_name", StringType(), False),
        StructField("gender", StringType(), False),
        StructField("address", StringType(), False),
        StructField("post_code", StringType(), False),
        StructField("email", StringType(), False),
        StructField("username", StringType(), False),
        StructField("registered_date", StringType(), False),
        StructField("phone", StringType(), False),
        StructField("picture", StringType(), False)
    ])

    df = spark_df.selectExpr("CAST(value AS STRING)") \
        .select(from_json(col('value'), schema).alias('data')).select("data.*")

    return df


if __name__ == "__main__":
    # create spark connection
    spark_conn = get_spark_connection()

    if spark_conn is not None:
        # connect to kafka with spark connection
        spark_df = get_data_from_kafka(spark_conn)
        parsed_df = parse_kafka_stream(spark_df)
        session = get_cassandra_connection()

        if session is not None:
            create_cassandra_keyspace(session)
            create_cassandra_table(session)

            print("[INFO] Starting Streaming")

            streaming_query = (parsed_df.writeStream.format("org.apache.spark.sql.cassandra")
                               .option('checkpointLocation', '/tmp/checkpoint')
                               .option('keyspace', 'spark_stream')
                               .option('table', 'users')
                               .start())

            streaming_query.awaitTermination()